+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Drivetrains/mecanumDriveRO_Arm.java
@@ -1,22 +1,18 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Drivetrains;
 
 //javadocs here: https://javadoc.io/doc/org.firstinspires.ftc
 //ftc docs here: https://ftc-docs.firstinspires.org/en/latest/programming_resources/index.html
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.IMU;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.hardware.Servo;
-
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 
 // Drive train controls for mecanum drive
 // Mecanum drive allows omnidirectional movement
 
-@TeleOp(name = "Mecanum - DO")
-public class mecanumDrive extends LinearOpMode {
+@TeleOp(name = "Mecanum - RO")
+public class mecanumDriveRO_Arm extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
         //create objects for motors
@@ -33,9 +29,9 @@ public class mecanumDrive extends LinearOpMode {
         Servo droneServo = hardwareMap.servo.get("droneServo");
 
         //reverse right side motors. reverse left side if goes backwards
-        frMotor.setDirection(DcMotorSimple.Direction.REVERSE); // IDK MAN, WE NEED TO TEST
+        frMotor.setDirection(DcMotorSimple.Direction.FORWARD); // IDK MAN, WE NEED TO TEST
         flMotor.setDirection(DcMotorSimple.Direction.FORWARD);
-        brMotor.setDirection(DcMotorSimple.Direction.REVERSE);
+        brMotor.setDirection(DcMotorSimple.Direction.FORWARD);
         blMotor.setDirection(DcMotorSimple.Direction.REVERSE);
 
         slideMotor.setDirection(DcMotorSimple.Direction.REVERSE);
@@ -46,25 +42,14 @@ public class mecanumDrive extends LinearOpMode {
         armRotateMotor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         slideMotor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
-        armRotateMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        slideMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        //armRotateMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        //slideMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-        armRotateMotor.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        //armRotateMotor.setTargetPosition(0);
 
+        //armRotateMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
-        //gets the IMU (Inertial Measurement Unit) from the hardware map
-        IMU imu = hardwareMap.get(IMU.class, "imu");
-
-        //sets orientation. change to match final robot
-        //default is Logo Up and USB Forward
-        IMU.Parameters parameters = new IMU.Parameters(new RevHubOrientationOnRobot(
-                RevHubOrientationOnRobot.LogoFacingDirection.RIGHT,
-                RevHubOrientationOnRobot.UsbFacingDirection.UP
-        ));
-
-        imu.initialize(parameters);
-
-        double servoPosition = 0.3;
+        //armRotateMotor.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
         boolean droneLaunched = false;
 
@@ -88,6 +73,7 @@ public class mecanumDrive extends LinearOpMode {
 
         //waits for start of game
         waitForStart();
+        droneServo.setPosition(0);
 
         boolean intakeButtonPressed = false;
 
@@ -101,91 +87,72 @@ public class mecanumDrive extends LinearOpMode {
             double rTrigger = gamepad1.right_trigger;
             double lTrigger = gamepad1.left_trigger;
 
-
-            //this button should be hard to hit on accident
-            //change if necessary
-            if (gamepad1.start) {
-                imu.resetYaw(); // reset the yaw of the robot (obvious)
-            }
-
-
-
-            //retrieves the yaw of the robot
-            double robotYaw = imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.RADIANS);
-            telemetry.addData("Robot Yaw",robotYaw);
-
-            //calculates how much the robot should turn
-            //directions are absolute
-            //ex. if the stick is moved up, the robot will always move away from the drive team.
-            double rotationX = stickX * Math.cos(-robotYaw) - stickY * Math.sin(-robotYaw);
-            double rotationY = stickX * Math.sin(-robotYaw) + stickY * Math.cos(-robotYaw);
-
-//            rotationX = rotationX * 1.1; //counteract imperfect strafing
-
-
             //in our case, our servos have a range of 300 degrees
             //the position numbers are a fraction of these 300 degrees
             // The position goes from 0 to 1
             // This represents the fraction of 300 degrees the motor should be at
             // eg. 0.5 would be 150 degrees & 0.1 would be 30.
             if (gamepad1.right_bumper){
-                servoPosition=0.5;
+                armTopServo.setPosition(0.6);
             } else if (gamepad1.left_bumper) {
-                servoPosition=0;
+                armTopServo.setPosition(0.125);
             }
-            telemetry.addData("Arm is up",(servoPosition>=0.4));
 
-            if (gamepad1.a && !intakeButtonPressed) { // Toggle intake motor on/off
-                intakeMotor.setPower(1-intakeMotor.getPower());
-                telemetry.addLine("Intake is moving");
-                intakeButtonPressed = true;
+            if (gamepad1.a) {
+                if (!intakeButtonPressed) {
+                    intakeMotor.setPower(1 - intakeMotor.getPower());
+                    telemetry.addLine("Intake is moving");
+                    intakeButtonPressed = true;
+                }
             } else {
                 telemetry.addLine("Intake is stopped");
+                intakeButtonPressed=false;
             }
 
 
-            // controls to rotate the whole arm up and down (forwards and backwards)
+
+            /// controls to rotate the whole arm up and down (forwards and backwards)
+            if (gamepad1.dpad_up) {
+                armRotateMotor.setPower(.7); //makes the arm motors rotate forwards slowly
+                telemetry.addLine("Moving arm forward");
+            } else if (gamepad1.dpad_down) {
+                armRotateMotor.setPower(-.7); //makes the arm motors rotate backwards slowly
+                telemetry.addLine("Moving arm backward");
+            } else {
+                armRotateMotor.setPower(0);
+                telemetry.addLine("Arm isn't moving");
+            }
+
+
             // only changes position when the motor isn't busy, (hopefully) making controls more precise
             // 5700.4 counts per revolution
-            if (gamepad1.dpad_up && !armRotateMotor.isBusy()) {
+            /*
+            if (gamepad1.dpad_up && !armRotateMotor.isBusy() ) {
                 armRotateMotor.setTargetPosition(armRotateMotor.getCurrentPosition() + 50); //makes the arm motors rotate forwards slowly
             } else if (gamepad1.dpad_down && !armRotateMotor.isBusy()) {
                 armRotateMotor.setTargetPosition(armRotateMotor.getCurrentPosition() - 50); //makes the arm motors rotate backwards slowly
             }// TODO: add telemetry
 
 
+             */
+            telemetry.addData("Arm Target Position", armRotateMotor.getTargetPosition());
+
             //moves the drone servo to the launch position
             if (gamepad1.back) {
-                droneServo.setPosition(0);
+                droneServo.setPosition(1);
             }
             telemetry.addData("Drone Launched",droneLaunched);
 
-            //denominator is either the motor power or 1, depending on which is larger.
-            //ensures all powers are the same ratio, but only when
-            //at least one power is <-1 or >1
-            double denominator = Math.max(Math.abs(rotationY) + Math.abs(rotationX) + Math.abs(rStickX), 1);
-            // Basically, it shrinks the movement amounts of each motor to fit in the range [-1,1]
-            // The motors cannot move more then 1 or less then -1, so it is necessary to shrink the
-            // movements amount to preserve the ratio between them all.
-            // If no values are above 1, then denominator is 1 (no effect)
+            double denominator = Math.max(Math.abs(stickX) + Math.abs(stickY) + Math.abs(rStickX), 1);
 
-            // calculate how much each motor should move`
-            double flPower = (rotationY + rotationX + rStickX) / denominator;
-            double frPower = (rotationY - rotationX - rStickX) / denominator;
-            double blPower = (rotationY - rotationX + rStickX) / denominator;
-            double brPower = (rotationY + rotationX - rStickX) / denominator;
-
-            flMotor.setPower(flPower); // move the motors based on calculations
-            frMotor.setPower(frPower);
-            blMotor.setPower(blPower);
-            brMotor.setPower(brPower);
+            flMotor.setPower((stickY + stickX + rStickX) / denominator);
+            frMotor.setPower((stickY - stickX - rStickX) / denominator);
+            blMotor.setPower((stickY - stickX + rStickX) / denominator);
+            brMotor.setPower((stickY + stickX - rStickX) / denominator);
 
             slideMotor.setPower(rTrigger-lTrigger+0.05); // move slide motor
             // the 0.05 is to counteract gravity
             // telemetry.addData("current arm motion:",rTrigger-lTrigger);
-
-            armTopServo.setPosition(servoPosition);
-            telemetry.addData("Arm Servo Position", servoPosition);
 
             telemetry.update();
         }
