+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Drivetrains/tankDrive.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Drivetrains;
 
 //javadocs here: https://javadoc.io/doc/org.firstinspires.ftc
 //ftc docs here: https://ftc-docs.firstinspires.org/en/latest/programming_resources/index.html
