+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Drivetrains/tankDriveArm.java
@@ -1,18 +1,23 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Drivetrains;
 
 //javadocs here: https://javadoc.io/doc/org.firstinspires.ftc
 //ftc docs here: https://ftc-docs.firstinspires.org/en/latest/programming_resources/index.html
+
+import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
+import com.qualcomm.robotcore.hardware.IMU;
 import com.qualcomm.robotcore.hardware.Servo;
+
+import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 
 // Drive train controls for mecanum drive
 // Mecanum drive allows omnidirectional movement
 
-@TeleOp(name = "Mecanum - RO")
-public class mecanumDriveRO_Arm extends LinearOpMode {
+@TeleOp(name = "Tank")
+public class tankDriveArm extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
         //create objects for motors
@@ -29,9 +34,9 @@ public class mecanumDriveRO_Arm extends LinearOpMode {
         Servo droneServo = hardwareMap.servo.get("droneServo");
 
         //reverse right side motors. reverse left side if goes backwards
-        frMotor.setDirection(DcMotorSimple.Direction.FORWARD); // IDK MAN, WE NEED TO TEST
+        frMotor.setDirection(DcMotorSimple.Direction.REVERSE); // IDK MAN, WE NEED TO TEST
         flMotor.setDirection(DcMotorSimple.Direction.FORWARD);
-        brMotor.setDirection(DcMotorSimple.Direction.FORWARD);
+        brMotor.setDirection(DcMotorSimple.Direction.REVERSE);
         blMotor.setDirection(DcMotorSimple.Direction.REVERSE);
 
         slideMotor.setDirection(DcMotorSimple.Direction.REVERSE);
@@ -42,14 +47,12 @@ public class mecanumDriveRO_Arm extends LinearOpMode {
         armRotateMotor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         slideMotor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
-        //armRotateMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        //slideMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+//        armRotateMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+//        slideMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-        //armRotateMotor.setTargetPosition(0);
+//        armRotateMotor.setTargetPosition(0);
 
-        //armRotateMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-        //armRotateMotor.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+//        armRotateMotor.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
         boolean droneLaunched = false;
 
@@ -57,8 +60,8 @@ public class mecanumDriveRO_Arm extends LinearOpMode {
         telemetry.addLine("Variables initialized");
         telemetry.addLine();
         telemetry.addLine("Controls:");
-        telemetry.addLine("Left Stick - Move (up, down, left, right)");
-        telemetry.addLine("Right Stick- Turn");
+        telemetry.addLine("Left Stick - Left Side Wheels");
+        telemetry.addLine("Right Stick - Right Side Wheels");
         telemetry.addLine("Start - Reset Yaw");
         telemetry.addLine("RT - Extend Arm");
         telemetry.addLine("LT - Retract Arm");
@@ -67,6 +70,7 @@ public class mecanumDriveRO_Arm extends LinearOpMode {
         telemetry.addLine("D-Pad Up - Rotate Arm Body Forwards");
         telemetry.addLine("D-Pad Down - Rotate Arm Body Backwards");
         telemetry.addLine("A button - Toggle intake on/off");
+        telemetry.addLine("Back - Launch Drone");
         telemetry.addLine();
         telemetry.addLine("Ready to start");
         telemetry.update();
@@ -79,9 +83,8 @@ public class mecanumDriveRO_Arm extends LinearOpMode {
 
         while (opModeIsActive()) {
             //gamepad variables
-            double stickY = -gamepad1.left_stick_y; //Y stick value is REVERSED
-            double stickX = gamepad1.left_stick_x;
-            double rStickX = gamepad1.right_stick_x;
+            double lstickY = -gamepad1.left_stick_y; //Y stick value is REVERSED
+            double rStickY = gamepad1.right_stick_y;
 
             //triggers for arm extending
             double rTrigger = gamepad1.right_trigger;
@@ -98,20 +101,16 @@ public class mecanumDriveRO_Arm extends LinearOpMode {
                 armTopServo.setPosition(0.125);
             }
 
-            if (gamepad1.a) {
-                if (!intakeButtonPressed) {
-                    intakeMotor.setPower(1 - intakeMotor.getPower());
-                    telemetry.addLine("Intake is moving");
-                    intakeButtonPressed = true;
-                }
+            if (gamepad1.a && !intakeButtonPressed) { // Toggle intake motor on/off
+                intakeMotor.setPower(1-intakeMotor.getPower());
+                telemetry.addLine("Intake is moving");
+                intakeButtonPressed = true;
             } else {
                 telemetry.addLine("Intake is stopped");
-                intakeButtonPressed=false;
             }
 
 
-
-            /// controls to rotate the whole arm up and down (forwards and backwards)
+            // controls to rotate the whole arm up and down (forwards and backwards)
             if (gamepad1.dpad_up) {
                 armRotateMotor.setPower(.7); //makes the arm motors rotate forwards slowly
                 telemetry.addLine("Moving arm forward");
@@ -119,23 +118,20 @@ public class mecanumDriveRO_Arm extends LinearOpMode {
                 armRotateMotor.setPower(-.7); //makes the arm motors rotate backwards slowly
                 telemetry.addLine("Moving arm backward");
             } else {
-                armRotateMotor.setPower(0);
                 telemetry.addLine("Arm isn't moving");
             }
 
-
             // only changes position when the motor isn't busy, (hopefully) making controls more precise
+
             // 5700.4 counts per revolution
             /*
-            if (gamepad1.dpad_up && !armRotateMotor.isBusy() ) {
+            if (gamepad1.dpad_up && !armRotateMotor.isBusy()) {
                 armRotateMotor.setTargetPosition(armRotateMotor.getCurrentPosition() + 50); //makes the arm motors rotate forwards slowly
             } else if (gamepad1.dpad_down && !armRotateMotor.isBusy()) {
                 armRotateMotor.setTargetPosition(armRotateMotor.getCurrentPosition() - 50); //makes the arm motors rotate backwards slowly
             }// TODO: add telemetry
-
-
              */
-            telemetry.addData("Arm Target Position", armRotateMotor.getTargetPosition());
+
 
             //moves the drone servo to the launch position
             if (gamepad1.back) {
@@ -143,17 +139,15 @@ public class mecanumDriveRO_Arm extends LinearOpMode {
             }
             telemetry.addData("Drone Launched",droneLaunched);
 
-            double denominator = Math.max(Math.abs(stickX) + Math.abs(stickY) + Math.abs(rStickX), 1);
+            flMotor.setPower(lstickY);
+            blMotor.setPower(lstickY);
 
-            flMotor.setPower((stickY + stickX + rStickX) / denominator);
-            frMotor.setPower((stickY - stickX - rStickX) / denominator);
-            blMotor.setPower((stickY - stickX + rStickX) / denominator);
-            brMotor.setPower((stickY + stickX - rStickX) / denominator);
+            frMotor.setPower(rStickY);
+            brMotor.setPower(rStickY);
 
             slideMotor.setPower(rTrigger-lTrigger+0.05); // move slide motor
             // the 0.05 is to counteract gravity
             // telemetry.addData("current arm motion:",rTrigger-lTrigger);
-
             telemetry.update();
         }
     }
