+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Examples/visionExample.java
@@ -1,33 +1,23 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Examples;
 
 import android.util.Size;
 
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.firstinspires.ftc.vision.VisionPortal;
 import org.firstinspires.ftc.vision.apriltag.AprilTagDetection;
 import org.firstinspires.ftc.vision.apriltag.AprilTagGameDatabase;
 import org.firstinspires.ftc.vision.apriltag.AprilTagProcessor;
+//import org.firstinspires.ftc.vision.tfod.TfodProcessor;
 
 import java.util.List;
 
-@Autonomous(name="Vision Driving - RED")
-public class visionDrivingExampleRed extends LinearOpMode {
+@TeleOp(name="Vision Example", group = "Examples")
+public class visionExample extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
-        int navigateTo = 5; //set the tag number to navigate towards
-
-        //create objects for motors
-        // f=front, b=back, l=left, r=right
-        DcMotor flMotor = hardwareMap.dcMotor.get("motorFL");
-        DcMotor frMotor = hardwareMap.dcMotor.get("motorFR");
-        DcMotor blMotor = hardwareMap.dcMotor.get("motorBL");
-        DcMotor brMotor = hardwareMap.dcMotor.get("motorBR");
-
-
         // Create apriltag processor
         AprilTagProcessor tagProcessor = new AprilTagProcessor.Builder()
                 .setDrawCubeProjection(true)
@@ -39,17 +29,13 @@ public class visionDrivingExampleRed extends LinearOpMode {
         // Add camera
         VisionPortal vPortal = new VisionPortal.Builder()
                 .setCamera(hardwareMap.get(WebcamName.class, "Webcam 1"))
-                .setCameraResolution(new Size(640,480))
+                .setCameraResolution(new Size(1920,1080))
                 .addProcessor(tagProcessor)
                 .enableLiveView(true)
                 .build();
-        // Used to view what the robot sees on screen during init
+        // Used to view what the robot sees on screen
 
         telemetry.addLine("Ready to start");
-        telemetry.addLine();
-        telemetry.addLine("Range is the distance to tag");
-        telemetry.addLine("Bearing and Elevation are the angles to the tag");
-        telemetry.addLine("idk man read the docs");
         telemetry.update();
 
         waitForStart();
@@ -67,13 +53,13 @@ public class visionDrivingExampleRed extends LinearOpMode {
             // process each tag in the list
             for (AprilTagDetection tagDetection : allTagDetections) {
 
-                tagIdCode = tagDetection.id; // Save current tag id to variable
-
-                if (tagDetection.metadata != null && tagIdCode == navigateTo) {
+                if (tagDetection.metadata != null) {
                     /*
                     Checks if current tag is not null
                     Check is not necessary when only reading the tag id code
                     */
+
+                    tagIdCode = tagDetection.id; // Save current tag id to variable
 
                     // place any code based on detected id here
                     // example code:
@@ -84,36 +70,16 @@ public class visionDrivingExampleRed extends LinearOpMode {
                     double tagBearing = tagDetection.ftcPose.bearing; // left/right
                     double tagElevation = tagDetection.ftcPose.elevation; // up/down
 
-                    if (tagBearing > 2) { //if the bearing is positive, turn counter-clockwise to the tag
-                        frMotor.setPower(1);
-                        flMotor.setPower(-1);
-                        brMotor.setPower(1);
-                        blMotor.setPower(-1);
-                    } else if (tagBearing < -2) { //if the bearing is negative, turn clockwise to the tag
-                        frMotor.setPower(-1);
-                        flMotor.setPower(1);
-                        brMotor.setPower(-1);
-                        blMotor.setPower(1);
-                    } else if (tagRange > 2){ //if neither is true, go straight. makes sure the robot doesn't ram into the tag
-                        frMotor.setPower(1);
-                        flMotor.setPower(1);
-                        brMotor.setPower(1);
-                        blMotor.setPower(1);
-                    }
-
                     telemetry.addData("Tag "+ tagIdCode +" Range:", tagRange);
                     telemetry.addData("Tag "+ tagIdCode +" Bearing:", tagBearing);
                     telemetry.addData("Tag "+ tagIdCode +" Elevation:", tagElevation);
-                } else{
-                    //turns right if no tags are detected
-                    frMotor.setPower(-1);
-                    flMotor.setPower(1);
-                    brMotor.setPower(-1);
-                    blMotor.setPower(1);
 
-                    telemetry.addLine("tag "+navigateTo+ " not detected");
                 }
             }
+            telemetry.addLine();
+            telemetry.addLine("Range is the distance to tag");
+            telemetry.addLine("Bearing and Elevation are the angles to the tag");
+            telemetry.addLine("idk man read the docs");
             telemetry.update();
         }
     }
