+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Misc/motorPortTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Misc;
 
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
